+++ packages/core-data/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Internal
+
+-   Remove unused dependency `@wordpress/is-shallow-equal` ([#82103](https://github.com/WordPress/gutenberg/pull/82103)).
+-   Use the `.jsx` extension for JavaScript source files that contain JSX ([#80990](https://github.com/WordPress/gutenberg/pull/80990)).
+-   Remove tsconfig project references to packages that are not dependencies ([#82106](https://github.com/WordPress/gutenberg/pull/82106)).
+
 ## 7.54.0 (2026-08-26)
 
 ### Bug Fixes